bug: Using intptr_t type instead of long to resolve a warning
There were a few places where pointers were being checked by casting to long. This should use intptr_t instead.
This type should be available on most, if not all current compilers. If it is not available, it is defined as long long for now. That can be further refined if/when a compiler does not have intptr_t and needs a better definition than this.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>