clean: Cleaning up unused variable warning when !JSON_LESS_MEMORY
Cleaning up a warning by adding a cast to void to stop an unused varable/parameter warning. This happens when JSON_LESS_MEMORY is not defined and the warning level is high enough. This effectively does nothing other than make the warning disappear. This was the easier solution to implement than the macros we have for this in opensea-common and did not want to reimplement that or add a lot of ifdefs.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>