feat: Making changes for C++17 compatibility
Making changes for C++17 compatibility. This is done by removing some places where dynamic exceptions were being defined which was allowed in C++98 and 03, but removed in 11. As of 17, this is treated as an error.

The best solution that seems to cause no impact is to comment this out in the functions where it was.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>